==> 03_rolle_energy.Rmd <==
---
title: "Energiebedarfs-Abschätzung der Gebäude in Rolle & Mont-Sur-Rolle"
author: "Curdin Derungs"
date: "April 2018"
output:
  html_document:
      toc: true
      toc_float: true
---

```{r setup, include=FALSE}

#loading libraries
library(sp)
library(htmlTable)
library(leaflet)
library(rgdal)

#empty workspace
rm(list=ls())
```


## Gebäudedaten aus Preprocessing einlesen

```{r readerGeb, warning=FALSE, paged.print=FALSE}
load("output/geom/rolle_buildings_with_gwr.Rdata")
#summary(rolle_bld_out@data)
```

## Funktion zur Abschätzung vom Energiebedarf von Gebäuden
In diesem Schritt wird von der Gebäudefläche * Anzahl Stockwerke und dem Gebäudealter der jährliche Energieverbrauch geschätzt. Die Schätzung beruht auf Daten aus dem [Energieplanungsbereicht](https://www.zh.ch/internet/de/aktuell/news/medienmitteilungen/2014/energieplanungsbericht_zeigt_erfolge_und_herausforderungen/_jcr_content/contentPar/downloadlist/downloaditems/44_1389359316861.spooler.download.1389359268717.pdf/Energieplanungsbericht-2013.pdf) des Kantons Zürich aus dem Jahr 2013 (S.19, Abbildung 14). Das Szenario "thomas" beruht auf Annahmen welche im Energiekonzept der Geimeinde Wohlen verwendet wurde (S.19, Abbildung 12).

```{r energieFunk, warning=FALSE, paged.print=FALSE}

#function to compute energie use per year from the age and the total area of a building
#different scenarios can be selected
#the conversion of age/are into energy use is based on a study cited in the header
energyUse<-function(age,area,scenario="2011"){
  if(scenario=="2011"){
    age.c<-as.numeric(as.character(
      cut(x = age,
          breaks=c(1000,1919,1959,1979,1989,1999,2020),
          labels=c(137,173,170,127,95,87))))
  } else if(scenario=="1990"){
    age.c<-as.numeric(as.character(
      cut(x = age,
          breaks=c(1000,1919,1959,1979,1989,1999,2020),
          labels=c(187,217,223,170,100,90))))
  }else if(scenario=="2006"){
    age.c<-as.numeric(as.character(
      cut(x = age,
          breaks=c(1000,1919,1959,1979,1989,1999,2020),
          labels=c(142,177,180,140,100,90))))
  }else if(scenario=="minergie"){
    age.c<-as.numeric(as.character(
      cut(x = age,
          breaks=c(1000,2001,2020),
          labels=c(60,38))))
  }else if(scenario=="thomas"){
    age.c<-as.numeric(as.character(
      cut(x = age,
          breaks=c(1000,1919,1959,1979,1989,1994,1999,2004,2009,2020),
          labels=c(140,178,171,120,95,88,80,70,50))))
  }else{
    stop("<scenario> should be: 1990, 2006, 2011, minergie or thomas")
  }
  return(age.c*area)
}

age<-as.numeric(rolle_bld_out$GBAUJ)

rolle_bld_out$area_polygon<-round(sapply(rolle_bld_out@polygons,function(x){x@area}),1)
area<-rolle_bld_out$area_polygon*as.numeric(rolle_bld_out$GASTW)
rolle_bld_out$ebf<-area

rolle_bld_out$energy_use<-energyUse(age,area,scenario = "thomas")
```

##Visualisierung

```{r geoviz, fig.width = 10, fig.height = 5.5}

rolle_bld_out.wgs<-spTransform(rolle_bld_out,CRSobj = CRS("+init=epsg:4326"))

qpal <- pal <- colorBin(
  palette = "YlGnBu",
  domain = log10(rolle_bld_out.wgs$energy_use),
  bins=10
)

#visualisierung mit leaflet
m <- leaflet() %>%
    
  addProviderTiles(providers$OpenStreetMap, group = "normal") %>%
  addProviderTiles(providers$Stamen.TonerLite, group = "OSM (b/w)") %>%
  addProviderTiles("Esri.WorldImagery", group = "ESRI") %>%
  
  addPolygons(data=rolle_bld_out.wgs,
              stroke = TRUE,
              fillOpacity = 0.8, 
              color = ~qpal(log10(rolle_bld_out.wgs$energy_use)),
              weight=1,
              popup=apply(rolle_bld_out.wgs@data[,c("GKAT_text","area_polygon","GBAUJ","GASTW","GHEIZ_text")], 1,  function(row) htmlTable(row, header = c("kat.","fl.","jahr","stockw","heiz"),align = "left")),
              #popup=paste(rolle_bld_out.wgs$EGID,rolle_bld_out.wgs$GKAT_text),
              group = "Haeuser")%>%
  
  addLayersControl(
    baseGroups = c("normal", "OSM (b/w)","ESRI"),
    overlayGroups = c("Haeuser"),
    options = layersControlOptions(collapsed = F)
  )%>%
  addLegend(pal = qpal, 
            values =log10(rolle_bld_out.wgs$energy_use), 
            opacity = 1,
            title = "log10(kWh/a)")

#leaflet karte ausführen
m
```

##Save all Information
```{r save}
#html karte speichern
wd<-getwd()
htmlwidgets::saveWidget(m, file=paste(wd,"/output/map/mapEnergy.html",sep=""),selfcontained = T)

#buildings speichern
save(rolle_bld_out,file="output/geom/rolle_buildings_with_energy.Rdata")
writeOGR(rolle_bld_out.wgs, "output/geom", "rolle_bld_out_energy.wgs", driver="ESRI Shapefile",overwrite_layer = T)
```
