==> 04_rolle_buildingPrototypes.Rmd <==
---
title: "rolle_buildingPrototypes"
author: "Curdin Derungs, ..."
date: "June 2018"
output:
  html_document:
      toc: true
      toc_float: true
---

```{r setup, include=FALSE}
#loading libraries
library(ggplot2)
library(rgeos)
library(gridExtra)
library(dplyr)

#empty workspace
rm(list=ls())
```


## Gebäudedaten mit Energieverbrauch laden

```{r readerGeb, warning=FALSE, paged.print=FALSE}
load("output/geom/rolle_buildings_with_energy.Rdata")
Sys.setlocale(category = "LC_ALL", locale = "en_US.UTF-8")
gwrCodes<-read.csv("data/gwrCodes.csv",encoding="UTF-8")
```


## Attribute zusammenlegen
...

```{r nachbarn, warning=FALSE, paged.print=FALSE}

summary(rolle_bld_out$GKLAS)

names(gwrCodes)[1]<-"id"
rolle_bld_out$GKLAS<-as.numeric(rolle_bld_out$GKLAS)
attrs<-left_join(rolle_bld_out@data[,c(1:12,25,26,27)], gwrCodes, by = c('GKLAS'='id'))
attrs$fullText<-NULL
names(attrs)[ncol(attrs)]<-"GKLAS_text"


attrs$GHEIZ<-as.numeric(attrs$GHEIZ)
attrs<-left_join(attrs, gwrCodes, by = c('GHEIZ'='id'))
attrs$fullText<-NULL
names(attrs)[ncol(attrs)]<-"GHEIZ_text"

attrs$GENHZ<-as.numeric(attrs$GENHZ)
attrs<-left_join(attrs, gwrCodes, by = c('GENHZ'='id'))
attrs$fullText<-NULL
names(attrs)[ncol(attrs)]<-"GENHZ_text"

attrs$GENWW<-as.numeric(attrs$GENWW)
attrs<-left_join(attrs, gwrCodes, by = c('GENWW'='id'))
attrs$fullText<-NULL
names(attrs)[ncol(attrs)]<-"GENWW_text"

attrs$GKAT<-as.numeric(attrs$GKAT)
attrs<-left_join(attrs, gwrCodes, by = c('GKAT'='id'))
attrs$fullText<-NULL
names(attrs)[ncol(attrs)]<-"GKAT_text"

rolle_bld_out@data<-attrs

```


## Attribute zusammenlegen
...

```{r nachbarn, warning=FALSE, paged.print=FALSE}

attrs$GBAUP<-cut(as.numeric(attrs$GBAUJ),breaks=c(0,1919,1945,1960,1970,1980,1985,1990,1995,2000,2005,2010,2015,2500))

test<-as.data.frame.table(table(attrs$GBAUP,factor(attrs$GKLAS_text)))
# library(reshape2)
# data_wide <- dcast(test, Var1 ~ Var2, value.var="Freq")
levels(test$Var2)

p1<-ggplot(test)+
  geom_bar(aes(x=Var1,y=Freq),stat="identity")+
  facet_grid(Var2~.)+
  theme(axis.text.x = element_text(angle = 90, hjust = 1,size=12))
```


##Visualisierung

```{r geoviz, fig.width = 10, fig.height = 5.5}

rolle_bld_out.wgs<-spTransform(rolle_bld_out,CRSobj = CRS("+init=epsg:4326"))

rolle_bld_centr.wgs = gCentroid(rolle_bld_out.wgs,byid=TRUE)


qpal <- pal <- colorBin(
  palette = "YlOrRd",
  domain = log10(rolle_bld_out.wgs$energy_use),
  bins=10
)

rolle_bld_out.wgs$GKLAS_text[is.na(rolle_bld_out.wgs$GKLAS_text)]<-"NotAvail"

factpal <- colorFactor(topo.colors(10), factor(rolle_bld_out.wgs$GKLAS_text))


#visualisierung mit leaflet
m <- leaflet() %>%
    
  addProviderTiles(providers$OpenStreetMap, group = "normal") %>%
  addProviderTiles(providers$Stamen.TonerLite, group = "OSM (b/w)") %>%
  addProviderTiles("Esri.WorldImagery", group = "ESRI") %>%
  
  addPolygons(data=rolle_bld_out.wgs,
              stroke = TRUE,
              fillOpacity = 0.8, 
              color = ~qpal(log10(rolle_bld_out.wgs$energy_use)),
              weight=1,
              #popup=paste(rolle_bld_out.wgs$EGID,rolle_bld_out.wgs$GKAT_text),
              group = "Haeuser")%>%
  
    addCircles(lng=rolle_bld_centr.wgs@coords[,1],
               lat=rolle_bld_centr.wgs@coords[,2],
               color = ~factpal(rolle_bld_out.wgs$GKLAS_text),
              group = "attrs")%>%
  
  addLayersControl(
    baseGroups = c("normal", "OSM (b/w)","ESRI"),
    overlayGroups = c("Haeuser","attrs"),
    options = layersControlOptions(collapsed = F)
  )%>%
  addLegend(pal = qpal, 
            values =log10(rolle_bld_out.wgs$energy_use), 
            opacity = 1,
            title = "log10(kWh/a)")

#leaflet karte ausführen
m
```
