==> 05_rolle_gridAggregation.Rmd <==
---
title: "Statistische Auswertung der Energiedaten von Rolle"
author: "Curdin Derungs, ..."
date: "April 2018"
output:
  html_document:
      toc: true
      toc_float: true
---

```{r setup, include=FALSE}
#loading libraries
library(raster)
library(sf)
library(rgeos)
library(plyr)
library(dplyr)
library(leaflet)
library(sp)

#empty workspace
rm(list=ls())
```


## Gebäudedaten mit Energieverbrauch laden

```{r readerGeb, warning=FALSE, paged.print=FALSE}
load("output/geom/rolle_buildings_with_energy.Rdata")
```


## Energiegrid Berechnen
...

```{r nachbarn, warning=FALSE, paged.print=FALSE}

pts<-gCentroid(rolle_bld_out,byid = T)
rolle_bld_pts<-SpatialPointsDataFrame(pts,rolle_bld_out@data)


e <- as(raster::extent(extent(rolle_bld_out)), "SpatialPolygons") %>% 
  st_as_sf()

grd.st <- st_make_grid(e, cellsize = c(100, 100)) %>%
  st_set_crs(21781)

grd.sp <- as(grd.st, "Spatial")

proj4string(grd.sp)<-CRS("+init=epsg:21781")
proj4string(rolle_bld_pts)<-CRS("+init=epsg:21781")

ov<-over(rolle_bld_pts,grd.sp)

ov<-data.frame(energ=rolle_bld_pts$energy_use,ov=ov)

ov<-ddply(ov,.(ov),summarise,sumEnergy=sum(energ))

grd.spdf<-SpatialPolygonsDataFrame(grd.sp,data.frame(ID=paste("ID",1:length(grd.sp),sep=""),iid=1:length(grd.sp)),match.ID = "ID")

dat<-left_join(grd.spdf@data, ov, by = c('iid'='ov'))
grd.spdf@data<-dat
```


## Energiegrid Visualisieren
...
```{r geoviz}
rolle_bld_out.wgs<-spTransform(rolle_bld_out,CRSobj = CRS("+init=epsg:4326"))
grd.spdf.wgs<-spTransform(grd.spdf,CRSobj = CRS("+init=epsg:4326"))


qpal <- pal <- colorBin(
  palette = "YlGnBu",
  domain = log10(rolle_bld_out.wgs$energy_use),
  bins=10
)

qpal.grid <- pal <- colorBin(
  palette = "BuPu",
  domain = grd.spdf.wgs$sumEnergy,
  bins=10
)

m <- leaflet() %>%
    
  addProviderTiles(providers$OpenStreetMap, group = "normal") %>%
  addProviderTiles(providers$Stamen.TonerLite, group = "OSM (b/w)") %>%
  addProviderTiles("Esri.WorldImagery", group = "ESRI") %>%
  
  addPolygons(data=rolle_bld_out.wgs,
              stroke = TRUE,
              fillOpacity = 0.8, 
              color = ~qpal(log10(rolle_bld_out.wgs$energy_use)),
              weight=1,
              group = "Haeuser")%>%
  
  addPolygons(data=grd.spdf.wgs,
              stroke = TRUE,
              fillOpacity = 0.6, 
              color = ~qpal.grid(grd.spdf.wgs$sumEnergy),
              weight=1,
              group = "Grid")%>%
  
  addLayersControl(
    baseGroups = c("normal", "OSM (b/w)","ESRI"),
    overlayGroups = c("Haeuser","Grid"),
    options = layersControlOptions(collapsed = F)
  )%>%
  
  addLegend(pal = qpal, 
            values =log10(rolle_bld_out.wgs$energy_use), 
            opacity = 1,
            title = "Haeuser - log10(kWh/a)",
            position = 'bottomright')%>%
  
  addLegend(pal = qpal.grid, 
            values =grd.spdf.wgs$sumEnergy, 
            opacity = 1,
            title = "Grid - kWh/a",
            position = 'bottomleft')

m
```

##Save all Information
```{r save}
#html karte speichern
wd<-getwd()
htmlwidgets::saveWidget(m, file=paste(wd,"/output/map/mapEnergyGrid.html",sep=""),selfcontained = T)

#grid speichern
save(grd.spdf,file="output/geom/rolle_energygrid.Rdata")
```
